move get_transaction_amount to external_api
